requirements.txt
annotated-types==0.7.0
anthropic==0.102.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
distro==1.9.0
docstring_parser==0.18.0
filelock==3.29.0
Flask==3.1.3
fsspec==2026.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.15
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
llvmlite==0.47.0
lxml==6.1.0
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
openai-whisper==20250625
packaging==26.2
pillow==12.2.0
pluggy==1.6.0
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
pytest==9.0.3
python-docx==1.2.0
python-dotenv==1.2.2
regex==2026.5.9
requests==2.34.2
setuptools==81.0.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.12.0
torch==2.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
Werkzeug==3.1.8
